Add photos, a Pixelfed instance
The deployment is particularly tricky to do, and relies on shared
storage for image uploads to work.

It follows the pattern of running two processes: one for Nginx, another
for PHP-FPM (which is provided by the custom Docker image). We share the
data between them so that Nginx can serve static assets and proxy PHP
requests through.

To allow uploaded images to be processed, the worker needs access to the
same filesystem. Unfortunately, this means that we also need to run
Horizon as a container in the same pod so that it has access to it.

Secrets rely on the 1Password Operator, which once defined are
mirrored from 1Password on an interval.

In the Nginx configuration, providing a custom path for `/storage` works
around a need to symlink the storage path by handling it directly with
Nginx.

The deployment is then exposed as a service, which sits behind another
Load Balancer.